# specs/settings-api-spec.yaml
openapi: 3.0.3
info:
  title: Settings API
  description: Enterprise-grade API for managing organizational settings, users, teams, permissions, audit logs, and postmortem templates.
  version: 1.0.0
servers:
  - url: https://api.ticketmanagement.internal/v1/settings
    description: Production Settings API
security:
  - BearerAuth: []
paths:
...
  /teams/{teamId}/members/{userId}:
    put:
      summary: Edit team member roles
      description: Updates the specific team-level roles granted to an existing member.
      tags:
        - Team Members
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTeamMemberRolesRequest'
      responses:
        '200':
          description: Member roles successfully updated.
        '400':
          $ref: '#/components/responses/BadRequestError'
          description: Warning logic triggers here if attempting to remove the mandatory 'Default User' access without reassignment.
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      summary: Remove member from team
      description: Ejects a member from the target team roster entirely.
      tags:
        - Team Members
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Successfully removed.
        '400':
          $ref: '#/components/responses/BadRequestError'
          description: Cannot remove the default team user without systemic re-assignment.

  # --------------------------------------------------------------------------------
  # Team Roles
  # --------------------------------------------------------------------------------
...
  /teams/{teamId}/roles/{roleId}:
    put:
      summary: Edit a custom team role
      description: Organically modifies the underlying boolean CRUD matrix of an existing custom role.
      tags:
        - Team Roles
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: roleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTeamRoleRequest'
      responses:
        '200':
          description: Matrix successfully updated.
        '400':
          $ref: '#/components/responses/BadRequestError'
        '403':
          description: Forbidden to edit system-immutable locked roles (e.g. "Manage Team").

    delete:
      summary: Delete a custom team role
      description: Obliterate an obsolete role explicitly from the active team hierarchy.
      tags:
        - Team Roles
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: roleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Correctly eliminated.
        '403':
          description: Forbidden to delete system-immutable roles.

  # --------------------------------------------------------------------------------
  # Postmortem Templates
  # --------------------------------------------------------------------------------
...
  /postmortem-templates/{id}:
    put:
      summary: Update template
      tags:
        - Postmortem Templates
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTemplateRequest'
      responses:
        '200':
          description: Success.

    delete:
      summary: Delete template
      tags:
        - Postmortem Templates
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted effectively.

  # --------------------------------------------------------------------------------
  # Audit Logs
  # --------------------------------------------------------------------------------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    StandardErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string

    # User Models
...
    UpdateTeamMemberRolesRequest:
      type: object
      required:
        - teamRoleIds
      properties:
        teamRoleIds:
          type: array
          items:
            type: string
            format: uuid

    # Team Role Models
    EntityPermissionMatrix:
      description: "A dynamic key-value matrix mapping system entities (e.g. 'schedules', 'postmortems') to their respective boolean CRUD capabilities."
      type: object
      additionalProperties:
        type: object
        properties:
          create:
            type: boolean
          read:
            type: boolean
          update:
            type: boolean
          delete:
            type: boolean
...
    UpdateTeamRoleRequest:
      type: object
      required:
        - permissions
      properties:
        name:
          type: string
        permissions:
          $ref: '#/components/schemas/EntityPermissionMatrix'

    # Postmortem Template Models
...
    CreateTemplateRequest:
      type: object
      required:
        - name
        - content
      properties:
        name:
          type: string
        content:
          type: string

    # Audit Log Models
...
  responses:
    BadRequestError:
      description: Bad Request Validation Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
    NotFoundError:
      description: Resource Not Found Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
...