# ticket-mgt-api-java/src/main/java/com/ticketmgt/settings/dto/StandardErrorResponse.java
package com.ticketmgt.settings.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class StandardErrorResponse {
    private String timestamp;
    private int status;
    private String error;
    private String message;
    private String path;
}