# ticket-mgt-api-java/src/main/java/com/ticketmgt/settings/dto/UpdateTeamRoleRequest.java
package com.ticketmgt.settings.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.Map;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class UpdateTeamRoleRequest {
    private String name;
    private Map<String, EntityPermissionMatrix> permissions;
}